requirements.txt
# Core LLM and API
langchain>=0.1.0
langchain-community>=0.0.20
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
aiohttp>=3.9.0
requests>=2.31.0

# Vector stores and embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.15
sentence-transformers>=2.2.2
torch>=2.1.0
transformers>=4.35.0

# Text processing
tiktoken>=0.5.1
pypdf>=3.17.0
python-multipart>=0.0.6

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.6.0

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=23.2.0
